Core files all compile again, and added argument builder support for arguments with a range, and enum support.

+++ src/main/java/com/laytonsmith/core/constructs/CArray.java
@@ -413,7 +413,7 @@ public class CArray extends Construct implements ArrayAccess{
         return mutVal;
     }
 
-    public long size() {
+    public int size() {
         if(associative_mode){
             return associative_array.size();
         } else {
